Handling forms with Formik and validating with Yup
Using useFormik hook for handling forms states, submissions, etc., instead of the custom hook "useControlledForm".
Pairing Formik with Yup for form validation on client side
The reason for choosing Formik and Yup instead of my custom hook and validation schema is the cleaner way of handling forms and better validation of inputs.